Do not cache the IP of the list server forever. Drop the ListServerReAddTime from 30 to 15 minutes to better handle lost connectivity with the list server. (Backported from r20670, r20671, and r20672)
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@21747 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/cURLManager.h
@@ -49,7 +49,7 @@ public:
   void setTimeCondition(timeCondition condition, time_t &t);
   void setInterface(const std::string interfaceIP);
   void setUserAgent(const std::string userAgent);
-  void setDNSCachingTime(int time);
+  void setDNSCachingTime(long time);
 
   void addFormData(const char *key, const char *value);
